refactor(ui-map): split ResourceLayerMap by responsibility
